PHOENIX-7314 Phoenix TTL for minor compaction

+++ phoenix-core-client/src/main/java/org/apache/phoenix/schema/DelegateTable.java
@@ -426,7 +426,7 @@ public class DelegateTable implements PTable {
     }
 
     @Override
-    public Long getMaxLookbackAge() {
+    public Integer getMaxLookbackAge() {
         return delegate.getMaxLookbackAge();
     }
